requirements.txt
streamlit
langchain
langchain-google-genai
chromadb
python-dotenv
langchain_community
unstructured